Move module options related components under componenst/module-options
